refactor: extract NestUpdatableEntity base class and simplify session manager
Deduplicate the identical __init__/session_manager and _async_update_objects
code from switch.py and select.py into a shared NestUpdatableEntity in
entity.py. Remove redundant credential storage from NestSessionManager —
it now reads issue_token/cookies/refresh_token directly from the client.
