+++ custom_components/nest_protect/select.py
@@ -9,8 +9,7 @@ from homeassistant.helpers.entity import EntityCategory
 
 from . import HomeAssistantNestProtectData
 from .const import DOMAIN, LOGGER
-from .entity import NestDescriptiveEntity
-from .session import NestSessionManager
+from .entity import NestUpdatableEntity
 
 
 @dataclass
@@ -61,22 +60,10 @@ async def async_setup_entry(hass, entry, async_add_devices):
     async_add_devices(entities)
 
 
-class NestProtectSelect(NestDescriptiveEntity, SelectEntity):
+class NestProtectSelect(NestUpdatableEntity, SelectEntity):
     """Representation of a Nest Protect Select."""
 
     entity_description: NestProtectSelectDescription
-
-    def __init__(
-        self,
-        bucket,
-        description,
-        areas,
-        client,
-        session_manager: NestSessionManager,
-    ) -> None:
-        """Initialize the select."""
-        super().__init__(bucket, description, areas, client)
-        self.session_manager = session_manager
 
     @property
     def current_option(self) -> str:
@@ -105,13 +92,3 @@ class NestProtectSelect(NestDescriptiveEntity, SelectEntity):
 
         result = await self._async_update_objects(objects)
         LOGGER.debug(result)
-
-    async def _async_update_objects(self, objects: list[dict]) -> dict:
-        """Update objects with automatic session refresh."""
-        await self.session_manager.ensure_session()
-        return await self.client.update_objects(
-            self.client.nest_session.access_token,
-            self.client.nest_session.userid,
-            self.client.transport_url,
-            objects,
-        )
